Move connection string parser into its own module
Extract the parser out of connection/mod.rs into connection/connection_string_parser.rs, leaving mod.rs as a thin module root that declares submodules and re-exports parse_connection_string. No behavior change.
